services/worker/requirements.txt
nats-py==2.6.0
redis==5.0.1
prometheus-client==0.19.0